Finish commons modularization
