+++ app/build.gradle
@@ -62,6 +62,7 @@ android {
 }
 
 dependencies {
+    implementation project(":commons")
     implementation project(":domain")
     implementation project(":data")
