+++ app/src/main/java/com/maximo/douglas/githubconsumer/di/OthersModule.kt
@@ -1,6 +1,6 @@
 package com.maximo.douglas.githubconsumer.di
 
-import com.maximo.douglas.githubconsumer.utils.CoroutineContextProvider
+import com.maximo.douglas.commons.utils.CoroutineContextProvider
 import org.koin.dsl.module
 
 val othersModule = module {
