+++ app/src/main/java/com/maximo/douglas/githubconsumer/ui/gitpullrequests/SelectGitPullRequestFragment.kt
@@ -15,17 +15,16 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import br.com.argmax.githubconsumer.R
 import br.com.argmax.githubconsumer.databinding.SelectGitPullRequestFragmentBinding
+import com.maximo.douglas.commons.utils.FragmentUtils.bundleContainsKeys
+import com.maximo.douglas.commons.utils.NavigationArgumentKeys.KEY_OWNER_LOGIN
+import com.maximo.douglas.commons.utils.NavigationArgumentKeys.KEY_REPOSITORY_NAME
+import com.maximo.douglas.commons.utils.StringUtils.gitPullRequestClosedLabelStringFormat
+import com.maximo.douglas.commons.utils.StringUtils.gitPullRequestOpenLabelStringFormat
 import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
 import com.maximo.douglas.domain.entities.gitpullrequest.PullRequestStateEnum
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel.SelectGitPullRequestViewModelState
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.adapters.SelectGitPullRequestAdapter
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.listeners.OnPullRequestClickListener
-import com.maximo.douglas.githubconsumer.utils.EndlessRecyclerOnScrollListener
-import com.maximo.douglas.githubconsumer.utils.FragmentUtils.bundleContainsKeys
-import com.maximo.douglas.githubconsumer.utils.NavigationArgumentKeys.KEY_OWNER_LOGIN
-import com.maximo.douglas.githubconsumer.utils.NavigationArgumentKeys.KEY_REPOSITORY_NAME
-import com.maximo.douglas.githubconsumer.utils.StringUtils.gitPullRequestClosedLabelStringFormat
-import com.maximo.douglas.githubconsumer.utils.StringUtils.gitPullRequestOpenLabelStringFormat
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class SelectGitPullRequestFragment : Fragment(), OnPullRequestClickListener {
@@ -101,7 +100,7 @@ class SelectGitPullRequestFragment : Fragment(), OnPullRequestClickListener {
 
         mBinding?.selectGitPullRequestFragmentRecyclerView?.adapter = mAdapter
         mBinding?.selectGitPullRequestFragmentRecyclerView?.addOnScrollListener(object :
-            EndlessRecyclerOnScrollListener() {
+            com.maximo.douglas.commons.utils.EndlessRecyclerOnScrollListener() {
             override fun onLoadMore() {
                 mApiRequestPage++
                 loadData()
