+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/ViewModelTestSuite.kt
@@ -12,4 +12,4 @@ import org.junit.runners.Suite
     SelectGitRepositoryViewModelSuiteTest::class,
     SelectGitPullRequestViewModelSuiteTest::class
 )
-class ViewModelSuiteTest
+class ViewModelTestSuite
