+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/gitrepository/context/SelectGitRepositoryViewModelTest.kt
@@ -3,7 +3,6 @@ package com.maximo.douglas.githubconsumer.viewmodels.gitrepository.context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import com.maximo.douglas.domain.entities.gitrepository.GitRepository
-import com.maximo.douglas.data.remote.gitrepository.GitRepositoryRemoteDataSource
 import com.maximo.douglas.githubconsumer.ui.gitrepositories.SelectGitRepositoryViewModel
 import com.maximo.douglas.githubconsumer.ui.gitrepositories.SelectGitRepositoryViewModel.SelectGitRepositoryViewModelState
 import com.maximo.douglas.githubconsumer.viewmodels.utils.TestContextProvider
