+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/utils/TestContextProvider.kt
@@ -1,10 +1,9 @@
 package com.maximo.douglas.githubconsumer.viewmodels.utils
 
-import com.maximo.douglas.githubconsumer.utils.CoroutineContextProvider
 import kotlinx.coroutines.Dispatchers
 import kotlin.coroutines.CoroutineContext
 
-class TestContextProvider : CoroutineContextProvider() {
+class TestContextProvider : com.maximo.douglas.commons.utils.CoroutineContextProvider() {
 
     override val Main: CoroutineContext = Dispatchers.Unconfined
     override val IO: CoroutineContext = Dispatchers.Unconfined
