+++ commons/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.maximo.douglas.domain">
+<manifest package="com.maximo.douglas.commons">
 
 </manifest>
