+++ commons/src/main/java/com/maximo/douglas/commons/utils/FragmentUtils.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.utils
+package com.maximo.douglas.commons.utils
 
 import android.os.Bundle
