+++ commons/src/main/java/com/maximo/douglas/commons/utils/ImageViewBindingAdapter.kt
@@ -1,10 +1,10 @@
-package com.maximo.douglas.githubconsumer.utils
+package com.maximo.douglas.commons.utils
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
-import br.com.argmax.githubconsumer.R
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CircleCrop
+import com.maximo.douglas.commons.R
 
 object ImageViewBindingAdapter {
