+++ commons/src/main/java/com/maximo/douglas/commons/utils/NavigationArgumentKeys.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.utils
+package com.maximo.douglas.commons.utils
 
 object NavigationArgumentKeys {
