+++ commons/src/main/java/com/maximo/douglas/commons/utils/ViewBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.utils
+package com.maximo.douglas.commons.utils
 
 import android.view.View
 import androidx.databinding.BindingAdapter
