+++ commons/src/main/res/drawable/ic_user.xml
