+++ commons/src/main/res/values/colors.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="colorPrimary">#37474F</color>
+    <color name="colorPrimaryDark">#263238</color>
+    <color name="colorAccent">#78909C</color>
+
+    <!--    Blue Gray Pallet    -->
+    <color name="blueGray50">#ECEFF1</color>
+    <color name="blueGray100">#CFD8DC</color>
+    <color name="blueGray200">#B0BEC5</color>
+    <color name="blueGray300">#90A4AE</color>
+    <color name="blueGray400">#78909C</color>
+    <color name="blueGray500">#607D8B</color>
+    <color name="blueGray600">#546E7A</color>
+    <color name="blueGray700">#455A64</color>
+    <color name="blueGray800">#37474F</color>
+    <color name="blueGray900">#263238</color>
+
+    <!--    Yellow Pallet    -->
+    <color name="yellow50">#FFFDE7</color>
+    <color name="yellow100">#FFF9C4</color>
+    <color name="yellow200">#FFF59D</color>
+    <color name="yellow300">#FFF176</color>
+    <color name="yellow400">#FFEE58</color>
+    <color name="yellow500">#FFEB3B</color>
+    <color name="yellow600">#FDD835</color>
+    <color name="yellow700">#FBC02D</color>
+    <color name="yellow800">#F9A825</color>
+    <color name="yellow900">#F57F17</color>
+
+    <!--    Blue Pallet    -->
+    <color name="blue50">#E1F5FE</color>
+    <color name="blue100">#B3E5FC</color>
+    <color name="blue200">#81D4FA</color>
+    <color name="blue300">#4FC3F7</color>
+    <color name="blue400">#29B6F6</color>
+    <color name="blue500">#03A9F4</color>
+    <color name="blue600">#039BE5</color>
+    <color name="blue700">#0288D1</color>
+    <color name="blue800">#0277BD</color>
+    <color name="blue900">#01579B</color>
+
+</resources>
