+++ commons/src/test/java/com/maximo/douglas/commons/UnitSuiteTest.kt
@@ -1,0 +1,11 @@
+package com.maximo.douglas.commons
+
+import com.maximo.douglas.commons.utils.StringUtilsTest
+import org.junit.runner.RunWith
+import org.junit.runners.Suite
+
+@RunWith(Suite::class)
+@Suite.SuiteClasses(
+    StringUtilsTest::class
+)
+class UnitSuiteTest
