+++ domain/src/main/java/com/maximo/douglas/domain/entities/gitpullrequest/GitPullRequest.kt
@@ -1,8 +1,8 @@
 package com.maximo.douglas.domain.entities.gitpullrequest
 
 import com.google.gson.annotations.SerializedName
+import com.maximo.douglas.commons.utils.StringUtils.compactStringWithDots
 import com.maximo.douglas.domain.entities.user.GitUser
-import com.maximo.douglas.domain.utils.StringUtils
 import java.io.Serializable
 
 data class GitPullRequest(
@@ -19,7 +19,7 @@ data class GitPullRequest(
 ) : Serializable {
 
     fun shortGitPullRequestBody(): String? {
-        return StringUtils.compactStringWithDots(body, 60)
+        return compactStringWithDots(body, 60)
     }
 
 }
