--- domain/src/main/java/com/maximo/douglas/domain/utils/StringUtils.kt
@@ -1,21 +1,0 @@
-package com.maximo.douglas.domain.utils
-
-object StringUtils {
-
-    fun compactStringWithDots(string: String?, maxLength: Int = 60): String? {
-        if ((string ?: "").length > maxLength) {
-            return string?.subSequence(0, maxLength) as String + "..."
-        }
-
-        return string
-    }
-
-    fun gitPullRequestOpenLabelStringFormat(quantity: Int? = 0): String {
-        return "$quantity open"
-    }
-
-    fun gitPullRequestClosedLabelStringFormat(quantity: Int? = 0): String {
-        return " / $quantity closed"
-    }
-
-}
