+++ domain/src/test/java/com/maximo/douglas/domain/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.maximo.douglas.domain
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
