+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':commons'
 include ':data'
 include ':domain'
 include ':app'
